feat: Add toggle for delete shortcut (Cmd+Backspace)
- Added "Enable delete shortcut" setting (disabled by default)
- Delete shortcut only works when explicitly enabled

+++ src/FolderFocusView.ts
@@ -252,7 +252,7 @@ export class FolderFocusView extends ItemView {
         this.enterOrOpen();
         break;
       case 'Backspace':
-        if (event.metaKey || event.ctrlKey) {
+        if ((event.metaKey || event.ctrlKey) && this.plugin.settings.enableDeleteShortcut) {
           event.preventDefault();
           this.deleteSelected();
         }
